+++ src/main/java/mediathek/config/Konstanten.java
@@ -26,6 +26,7 @@ import java.util.concurrent.TimeUnit;
 public class Konstanten {
     public static String ZAPP_API_URL = "https://api.zapp.mediathekview.de/";
     public static final String NEW_SENDER_ACTIVATED_QUESTION_CONFIG_KEY = "newSendersActivated.fourteen.three";
+    public static final String NEW_FILMLENGTH_ACTIVATED_QUESTION_CONFIG_KEY = "newFilmlengthActivated.fourteen.three";
     public static final long MINIMUM_MEMORY_THRESHOLD = 768 * FileUtils.ONE_MB;
     public static final Version MVVERSION = new Version(14, 4, 0);